requirements.txt
pyautogen~=0.2.0
azure-search-documents
azure-identity
python-dotenv
openai
